+++ driver-core/src/main/com/mongodb/operation/AsyncChangeStreamBatchCursor.java
@@ -19,8 +19,6 @@ package com.mongodb.operation;
 import com.mongodb.MongoChangeStreamException;
 import com.mongodb.MongoCursorNotFoundException;
 import com.mongodb.MongoNotPrimaryException;
-import com.mongodb.MongoQueryException;
-import com.mongodb.MongoSocketException;
 import com.mongodb.MongoSocketReadException;
 import com.mongodb.async.AsyncBatchCursor;
 import com.mongodb.async.SingleResultCallback;